Day 17 pt 1
Since the curve of the probe always lands back on y=0, the highest y
velocity must be the one which reaches the bottom of the area in one
step, which is conveniently the lower (first) y value. The x value does
not matter since we can choose some x velocity which ends up as 0 right
above a spot in the target area, meaning the probe falls straight down.