Day 20 pt 1
Put day 19 off for the longest time, and decided to skip it for now and
move on to day 20. Took a while to figure out that on every odd run with
the big input, there will be an infinite number of on pixels, which does
not happen in the smaller sample input.